# back/docs/openapi3.yaml
swagger: "2.0"
info:
  description: "Une application pour effectuer des relevés médicaux, les analyser et suivre sa santé."
  version: "1.0.0"
  title: "Mon assistance médical"
...
schemes:
- "https"
- "http"
paths:
...
  /utilisateur:
    post:
      tags:
      - "Utilisateur"
      summary: "Création d'un utilisateur"
      description: ""
      operationId: "addUser"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Utilisateur que l'on veut ajouter à la table."
        required: true
        schema:
          $ref: "#/definitions/Utilisateur"
      responses:
        "200":
          description: "Add OK"
        "500":
          description: "Add NOK"
#      security:
#      - petstore_auth:
#        - "write:pets"
#        - "read:pets"
    put:
      tags:
      - "Utilisateur"
      summary: "Modification d'un utilisateur"
      description: ""
      operationId: "updateUser"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Utilisateur que l'on veut modifier."
        required: true
        schema:
          $ref: "#/definitions/Utilisateur"
      responses:
        "200":
          description: "Update OK"
        "500":
          description: "Update NOK"
  /utilisateur/authentification:
    get:
      tags:
      - "Utilisateur"
      summary: "Création d'un utilisateur"
      description: ""
      operationId: "loginUser"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Utilisateur que l'on veut ajouter à la table."
        required: true
        schema:
          $ref: "#/definitions/UtilisateurLogin"
      responses:
        "200":
          description: "Utilisateur. Null si pas d'utilisateur"
          schema:
            $ref: "#/definitions/UtilisateurComplet"
        "500":
          description: "Login NOK"
  /utilisateur/{id}:
    get:
      tags:
      - "Utilisateur"
      summary: "Récupération d'un utilisateur grâce à son identifiant."
      description: ""
      operationId: "getUserById"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "id de l'utilisateur que l'on cherche"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Utilisateur et ses profils."
          schema:
            $ref: "#/definitions/UtilisateurComplet"
        "500":
          description: "Utilisateur non trouvé."
    delete:
      tags:
      - "Utilisateur"
      summary: "Suppression d'un utilisateur au moyen de son identifiant."
      description: ""
      operationId: "deleteUserById"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "id de l'utilisateur que l'on veut supprimer."
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Suppression OK."
        "500":
          description: "Suppression NOK."
  /utilisateur/{id}/profil/{idprofil}:
    post:
      tags:
      - "Utilisateur"
      summary: "Ajout d'un profil à un utilisateur."
      description: ""
      operationId: "addProfilToUser"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "id de l'utilisateur auquel on veut ajouter un profil."
        required: true
        type: "integer"
        format: "int64"
      - name: "idprofil"
        in: "path"
        description: "id du profil que l'on veut ajouter à l'utilisateur."
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Ajout OK."
        "500":
          description: "Ajout NOK."
    delete:
      tags:
      - "Utilisateur"
      summary: "Supression d'un profil à un utilisateur."
      description: ""
      operationId: "supprProfilToUser"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "id de l'utilisateur duquel on veut supprimer un profil."
        required: true
        type: "integer"
        format: "int64"
      - name: "idprofil"
        in: "path"
        description: "id du profil que l'on veut supprimer de l'utilisateur."
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Suppression OK."
        "500":
          description: "Suppression NOK."
  /utilisateur/{id}/releve:
    get:
      tags:
      - "Utilisateur"
      summary: "Récupération des relevés d'un utilisateur"
      description: "Récupération des relevés d'un utilisateur, possibilité de filtrer par type de relevé et de paginer la récupération par page de 5"
      operationId: "getUserReleve"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "id de l'utilisateur que l'on cherche"
        required: true
        type: "integer"
        format: "int64"
      - name: "page"
        in: "query"
        description: "numero de la page que l'on veut récupérer"
        required: false
        type: "integer"
        format: "int64"
      - name: "type"
        in: "query"
        description: "type des relevés que l'on veut récupérer."
        required: false
        type: "string"
        format: "int64"
      responses:
        "200":
          description: "Relevés trouvés."
          schema:
            $ref: "#/definitions/ListeRelevesMedicaux"
        "500":
          description: "Relevés non trouvés."


########################################
#                                      #
#         Requêtes pour profil         #
#                                      #
########################################
  /profil:
    post:
      tags:
      - "Profil"
      summary: "Création d'un profil utilisateur"
      description: ""
      operationId: "addProfil"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Profil d'utilisateur que l'on veut ajouter à la table."
        required: true
        schema:
          $ref: "#/definitions/Profil"
      responses:
        "200":
          description: "Profil"
          schema:
            $ref: "#/definitions/Profil"
        "500":
          description: "erreur"
    get:
      tags:
      - "Profil"
      summary: "Récupération des profils utilisateur"
      description: ""
      operationId: "getProfils"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        "200":
          description: "Liste des profils"
          schema:
            $ref: "#/definitions/Profils"
        "500":
          description: "erreur"
...
  /type_releve:
    get:
      tags:
      - "Type de relevé"
      summary: "Récupération des types de relevés médicaux."
      description: ""
      operationId: "getTypeReleve"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      responses:
        "200":
          description: "Récupération OK"
          schema:
            $ref: "#/definitions/TypeReleve"
        "500":
          description: "Récupération NOK"
    post:
      tags:
      - "Type de relevé"
      summary: "Ajout d'un type de relevé médical."
      description: ""
      operationId: "addTypeReleve"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Utilisateur que l'on veut ajouter à la table."
        required: true
        schema:
          $ref: "#/definitions/TypeReleve"
      responses:
        "200":
          description: "Ajout OK."
        "500":
          description: "Ajout NOK"
  

########################################
#                                      #
#         Requêtes pour analyse        #
#                                      #
########################################
...
  /analyse/relevé/{id}:
    get:
      tags:
      - "Analyse"
      summary: "Récupération des analyses d'un relevé"
      description: ""
      operationId: "getAnalyseByReleve"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifiant du relevé à analyser."
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Récupération OK"
          schema:
            $ref: "#/definitions/ListeAnalysesDangerositeCouleur"
        "500":
          description: "Récupération NOK"
          
########################################
#                                      #
#       Requêtes pour dangerosité      #
#                                      #
########################################
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Utilisateur:
    type: "object"
    required:
    - "id"
    - "nom"
    - "prenom"
    - "login"
    - "mot_de_passe"
    - "naissance"
    properties:
      id:
        type: "integer"
        format: "int64"
      nom:
        type: "string"
        example: "Fraysse"
      prenom:
        type: "string"
        example: "Marcel"
      login:
        type: "string"
        example: "marcel.fraysse"
      mot_de_passe:
        type: "string"
        example: "Mfraysse12"
      naissance:
        type: "string"
        example: "1997-04-23T00:00:00.511Z"
  UtilisateurComplet:
    type: "object"
    required:
    - "id"
    - "nom"
    - "prenom"
    - "login"
    - "mot_de_passe"
    - "naissance"
    properties:
      id:
        type: "integer"
        format: "int64"
      nom:
        type: "string"
        example: "Fraysse"
      prenom:
        type: "string"
        example: "Marcel"
      login:
        type: "string"
        example: "marcel.fraysse"
      mot_de_passe:
        type: "string"
        example: "Mfraysse12"
      naissance:
        type: "string"
        example: "1997-04-23T00:00:00.511Z"
      Profils:
        type: "array"
        items:
          type: "object"
          properties:
            schema:
              $ref: "#/definitions/Profil"
#    xml:
#      name: "Pet"
  UtilisateurLogin:
    type: "object"
    required:
    - "login"
    - "mot_de_passe"
    properties:
      login:
        type: "string"
        example: "marcel.fraysse"
      mot_de_passe:
        type: "string"
        example: "Mfraysse12"
  Profil:
    type: "object"
    required:
    - "label"
    properties:
      id:
        type: "integer"
        format: "int64"
      label:
        type: "string"
        example: "Adulte"
  Profils:
    type: "array"
    items:
      type: "object"
      properties:
        schema:
          $ref: "#/definitions/Profil"
  ReleveMedical:
    type: "object"
    required:
    - "id"
    - "prise_de_mesure"
    - "valeur"
    - "utilisateurId"
    - "typeReleveId"
    properties:
      id:
        type: "integer"
        example: 1
      prise_de_mesure:
        type: "string"
        example: "1997-04-23T00:00:00.511Z"
      valeur:
        type: "integer"
        example: 110
      utilisateurId:
        type: "integer"
        example: 19
      typeReleveId:
        type: "integer"
        example: 6
  ListeRelevesMedicaux:
    type: "array"
    items:
      type: "object"
      properties:
        schema:
          $ref: "#/definitions/ReleveMedical"
  Analyse:
    type: "object"
    required:
    - "id"
    - "mini"
    - "maxi"
    - "dangerositeId"
    - "profilId"
    - "typeReleveId"
    properties:
      id:
        type: "integer"
        example: 12
      mini:
        type: "integer"
        example: 95
      maxi:
        type: "integer"
        example: 110
      dangerositeId:
        type: "integer"
        example: 7
      profilId:
        type: "integer"
        example: 2
      typeReleveId:
        type: "integer"
        example: 3
...
  TypeReleve:
    type: "object"
    required:
    - "label"
    properties:
      id:
        type: "integer"
        format: "int64"
      label:
        type: "string"
        example: "Taille"
...
  DangerositeCouleur:
    type: "object"
    required:
    - "id"
    - "message"
    - "couleur"
    properties:
      id:
        type: "integer"
        example: 8
      message:
        type: "string"
        example: "A surveiller. Consulter votre médecin."
      couleur:
        type: "object"
        properties:
          schema:
            $ref: "#/definitions/Analyse"
  AnalyseDangerositeCouleur:
    type: "object"
    required:
    - "id"
    - "mini"
    - "maxi"
    - "dangerosite"
    - "profilId"
    - "typeReleveId"
    properties:
      id:
        type: "integer"
        example: 12
      mini:
        type: "integer"
        example: 95
      maxi:
        type: "integer"
        example: 110
      dangerosite:
        type: "object"
        properties:
          schema:
            $ref: "#/definitions/DangerositeCouleur"
      profilId:
        type: "integer"
        example: 2
      typeReleveId:
        type: "integer"
        example: 3
  ListeAnalysesDangerositeCouleur:
    type: "array"
    items:
      type: "object"
      properties:
        schema:
          $ref: "#/definitions/AnalyseDangerositeCouleur"